+++ src/sll/stack.c
@@ -13,8 +13,7 @@ sll_assembly_instruction_t* _acquire_next_instruction(sll_assembly_data_t* a_dt)
 	a_dt->_s.c--;
 	a_dt->_s.p++;
 	if (!a_dt->_s.c){
-		sll_page_size_t sz=sll_platform_get_page_size()*ASSEMBLY_INSTRUCTION_STACK_PAGE_ALLOC_COUNT;
-		void* n=sll_platform_allocate_page(sz);
+		void* n=sll_platform_allocate_page(SLL_ROUND_PAGE(ASSEMBLY_INSTRUCTION_STACK_ALLOC_SIZE),0);
 		*((void**)(a_dt->_s.e))=n;
 		*((void**)n)=NULL;
 		sll_assembly_instruction_t* s=(sll_assembly_instruction_t*)((char*)n+sizeof(void*));
@@ -22,7 +21,7 @@ sll_assembly_instruction_t* _acquire_next_instruction(sll_assembly_data_t* a_dt)
 		s->dt._p=a_dt->_s.p-1;
 		SLL_ASSERT(a_dt->_s.p->t==ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK);
 		a_dt->_s.p->dt._p=s+1;
-		a_dt->_s.c=(uint32_t)(((sz-sizeof(void*)-sizeof(sll_assembly_instruction_t)*2)/sizeof(sll_assembly_instruction_t)));
+		a_dt->_s.c=(uint32_t)(((SLL_ROUND_PAGE(ASSEMBLY_INSTRUCTION_STACK_ALLOC_SIZE)-sizeof(void*)-sizeof(sll_assembly_instruction_t)*2)/sizeof(sll_assembly_instruction_t)));
 		a_dt->_s.p=s+1;
 		s+=a_dt->_s.c+1;
 		s->t=ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK;
@@ -40,8 +39,7 @@ sll_node_t* _acquire_next_node(sll_compilation_data_t* c_dt){
 	c_dt->_s.c--;
 	c_dt->_s.p++;
 	if (!c_dt->_s.c){
-		sll_page_size_t sz=sll_platform_get_page_size()*OBJECT_STACK_PAGE_ALLOC_COUNT;
-		void* n=sll_platform_allocate_page(sz);
+		void* n=sll_platform_allocate_page(SLL_ROUND_PAGE(NODE_STACK_ALLOC_SIZE),0);
 		*((void**)(c_dt->_s.e))=n;
 		*((void**)n)=NULL;
 		sll_node_t* s=(sll_node_t*)((char*)n+sizeof(void*));
@@ -49,7 +47,7 @@ sll_node_t* _acquire_next_node(sll_compilation_data_t* c_dt){
 		s->dt._p=c_dt->_s.p-1;
 		SLL_ASSERT(c_dt->_s.p->t==NODE_TYPE_CHANGE_STACK);
 		c_dt->_s.p->dt._p=s+1;
-		c_dt->_s.c=(uint32_t)(((sz-sizeof(void*)-sizeof(sll_node_t)*2)/sizeof(sll_node_t)));
+		c_dt->_s.c=(uint32_t)(((SLL_ROUND_PAGE(NODE_STACK_ALLOC_SIZE)-sizeof(void*)-sizeof(sll_node_t)*2)/sizeof(sll_node_t)));
 		c_dt->_s.p=s+1;
 		s+=c_dt->_s.c+1;
 		s->t=NODE_TYPE_CHANGE_STACK;
@@ -62,8 +60,7 @@ sll_node_t* _acquire_next_node(sll_compilation_data_t* c_dt){
 
 
 sll_assembly_instruction_t* _get_instruction_at_offset(const sll_assembly_data_t* a_dt,sll_instruction_index_t off){
-	sll_page_size_t sz=sll_platform_get_page_size()*ASSEMBLY_INSTRUCTION_STACK_PAGE_ALLOC_COUNT;
-	sll_instruction_index_t cnt=(sll_instruction_index_t)(((sz-sizeof(void*)-sizeof(sll_node_t)*2)/sizeof(sll_node_t)));
+	sll_instruction_index_t cnt=(sll_instruction_index_t)(((SLL_ROUND_PAGE(ASSEMBLY_INSTRUCTION_STACK_ALLOC_SIZE)-sizeof(void*)-sizeof(sll_node_t)*2)/sizeof(sll_node_t)));
 	void* pg=a_dt->_s.s;
 	while (off>=cnt){
 		pg=*((void**)pg);
@@ -75,8 +72,7 @@ sll_assembly_instruction_t* _get_instruction_at_offset(const sll_assembly_data_t
 
 
 sll_node_t* _get_node_at_offset(const sll_compilation_data_t* c_dt,sll_node_offset_t off){
-	sll_page_size_t sz=sll_platform_get_page_size()*OBJECT_STACK_PAGE_ALLOC_COUNT;
-	sll_node_offset_t cnt=(sll_node_offset_t)(((sz-sizeof(void*)-sizeof(sll_node_t)*2)/sizeof(sll_node_t)));
+	sll_node_offset_t cnt=(sll_node_offset_t)(((SLL_ROUND_PAGE(NODE_STACK_ALLOC_SIZE)-sizeof(void*)-sizeof(sll_node_t)*2)/sizeof(sll_node_t)));
 	void* pg=c_dt->_s.s;
 	while (off>=cnt){
 		pg=*((void**)pg);
@@ -88,14 +84,13 @@ sll_node_t* _get_node_at_offset(const sll_compilation_data_t* c_dt,sll_node_offs
 
 
 void _init_assembly_stack(sll_assembly_data_t* a_dt){
-	sll_page_size_t sz=sll_platform_get_page_size()*ASSEMBLY_INSTRUCTION_STACK_PAGE_ALLOC_COUNT;
-	a_dt->_s.s=sll_platform_allocate_page(sz);
+	a_dt->_s.s=sll_platform_allocate_page(SLL_ROUND_PAGE(ASSEMBLY_INSTRUCTION_STACK_ALLOC_SIZE),0);
 	a_dt->_s.e=a_dt->_s.s;
 	*((void**)(a_dt->_s.s))=NULL;
 	sll_assembly_instruction_t* s=(sll_assembly_instruction_t*)((char*)(a_dt->_s.s)+sizeof(void*));
 	s->t=ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK;
 	s->dt._p=NULL;
-	a_dt->_s.c=(uint32_t)(((sz-sizeof(void*)-sizeof(sll_assembly_instruction_t)*2)/sizeof(sll_assembly_instruction_t)));
+	a_dt->_s.c=(uint32_t)(((SLL_ROUND_PAGE(ASSEMBLY_INSTRUCTION_STACK_ALLOC_SIZE)-sizeof(void*)-sizeof(sll_assembly_instruction_t)*2)/sizeof(sll_assembly_instruction_t)));
 	a_dt->_s.p=s+1;
 	s+=a_dt->_s.c+1;
 	s->t=ASSEMBLY_INSTRUCTION_TYPE_CHANGE_STACK;
@@ -105,14 +100,13 @@ void _init_assembly_stack(sll_assembly_data_t* a_dt){
 
 
 void _init_node_stack(sll_compilation_data_t* c_dt){
-	sll_page_size_t sz=sll_platform_get_page_size()*OBJECT_STACK_PAGE_ALLOC_COUNT;
-	c_dt->_s.s=sll_platform_allocate_page(sz);
+	c_dt->_s.s=sll_platform_allocate_page(SLL_ROUND_PAGE(NODE_STACK_ALLOC_SIZE),0);
 	c_dt->_s.e=c_dt->_s.s;
 	*((void**)(c_dt->_s.s))=NULL;
 	sll_node_t* s=(sll_node_t*)((char*)(c_dt->_s.s)+sizeof(void*));
 	s->t=NODE_TYPE_CHANGE_STACK;
 	s->dt._p=NULL;
-	c_dt->_s.c=(uint32_t)(((sz-sizeof(void*)-sizeof(sll_node_t)*2)/sizeof(sll_node_t)));
+	c_dt->_s.c=(uint32_t)(((SLL_ROUND_PAGE(NODE_STACK_ALLOC_SIZE)-sizeof(void*)-sizeof(sll_node_t)*2)/sizeof(sll_node_t)));
 	c_dt->_s.p=s+1;
 	s+=c_dt->_s.c+1;
 	s->t=NODE_TYPE_CHANGE_STACK;