+++ CHANGELOG.md
@@ -7,10 +7,17 @@
 - Default file API: `sll_api_file_std_handle` and `file_std_handle`
 - Handle clone API: `sll_handle_clone` ([#101])
 - Handles are now cloned when assigned to custom type objects ([#101])
+- Large page support: `SLL_ROUND_PAGE` and `SLL_ROUND_LARGE_PAGE` ([#102])
 
 ### Changed
 
 - Converting handels to strings now uses a new template
+- Memory page size can now be accessed via `SLL_PAGE_SIZE` and `SLL_LARGE_PAGE_SIZE` macros ([#102])
+- `sll_platform_allocate_page` now accepts an argument to allocate large pages ([#102])
+
+### Removed
+
+- Page size API: `sll_platform_get_page_size`
 
 ## [0.6.30] - 2021-12-05
 
@@ -1114,6 +1121,7 @@ Unfortunately, no versions were assigned before 2021-6-15 (:disappointed:), so t
 [0.1.4]: https://github.com/sl-lang/sll/compare/v0.1.3...v0.1.4
 [0.1.3]: https://github.com/sl-lang/sll/compare/v0.1.2...v0.1.3
 [0.1.2]: https://github.com/sl-lang/sll/tree/v0.1.2
+[#102]: https://github.com/sl-lang/sll/issues/102
 [#101]: https://github.com/sl-lang/sll/issues/101
 [#100]: https://github.com/sl-lang/sll/issues/100
 [#99]: https://github.com/sl-lang/sll/issues/99